feat(data): implement SQLCipher for database encryption and update ID types to Long
- Add SQLCipher for Room database encryption with `DisposableKeySupportFactory`.
- Update `genreId` and `movieId` types from `Int` to `Long` across entities, DAOs, DTOs, and navigation.
- Refactor `MovieRemoteMediator` to use `Long` for paging keys and fix genre-specific data clearing.
- Optimize `MovieListScreen` with item keys and content types for better list performance.
- Improve `MovieItem` UI with updated styling, layout, and image loading optimizations.
- Update dependencies for SQLCipher and Jetpack Security in `libs.versions.toml`.

+++ app/src/main/java/com/ratbyansa/moviedb/ui/viewmodel/MovieViewModel.kt
@@ -16,7 +16,7 @@ import kotlinx.coroutines.launch
 
 class MovieViewModel(private val repository: MovieRepository) : ViewModel() {
 
-    private val _currentGenreId = MutableStateFlow<Int?>(null)
+    private val _currentGenreId = MutableStateFlow<Long?>(null)
 
     @OptIn(ExperimentalCoroutinesApi::class)
     val moviePagingData: Flow<PagingData<MovieEntity>> = _currentGenreId
@@ -26,7 +26,7 @@ class MovieViewModel(private val repository: MovieRepository) : ViewModel() {
         }
         .cachedIn(viewModelScope)
 
-    fun setGenre(genreId: Int) {
+    fun setGenre(genreId: Long) {
         if (_currentGenreId.value == genreId) return
         _currentGenreId.value = genreId
     }
